long to int

+++ src/main/java/com/esri/core/geometry/OperatorImportFromGeoJsonCursor.java
@@ -40,7 +40,7 @@ public class OperatorImportFromGeoJsonCursor extends MapGeometryCursor {
 	}
 
 	@Override
-	public long getGeometryID() {
+	public int getGeometryID() {
 		return m_jsonStringCursor.getID();
 	}
